+++ pikaia/models/pikaia_model.py
@@ -248,6 +248,7 @@ class PikaiaModel(GeneticModel):
             "initial_org_fitness_range": self._initial_org_fitness_range,
             "org_similarity": self._org_similarity,
             "gene_similarity": self._gene_similarity,
+            "normalizations": self._strategy_normalizations,
             "y": self._y,
         }