+++ pikaia/schemas/__init__.py
@@ -1,6 +1,21 @@
 """Enums and schema definitions for pikaia."""
 
 from .preprocessing import FeatureType
-from .strategies import GeneStrategyEnum, MixStrategyEnum, OrgStrategyEnum
+from .strategies import (
+    GeneStrategyEnum,
+    MixStrategyEnum,
+    OrgStrategyEnum,
+    StrategyFormulation,
+    StrategyFormulationConfig,
+    StrategyNormalizations,
+)
 
-__all__ = ["GeneStrategyEnum", "OrgStrategyEnum", "MixStrategyEnum", "FeatureType"]
+__all__ = [
+    "StrategyFormulation",
+    "StrategyFormulationConfig",
+    "StrategyNormalizations",
+    "GeneStrategyEnum",
+    "OrgStrategyEnum",
+    "MixStrategyEnum",
+    "FeatureType",
+]